Move imgen to Yliss
It belongs there because I plan to eventually use it in Yliss.

+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: all init
+.PHONY: all init copy-assets pull update
 
 ###################
 # HTML generation #
@@ -8,12 +8,21 @@ all: draw/index.html draw/favicon.html draw/imgen.html
 %.html: %.org
 	./script/org-to-html.bash "$^" "$@"
 
-###############################
-# Dependencies initialisation #
-###############################
+##############
+# Submodules #
+##############
 init:
 	git submodule init
 	git submodule update
-	rmdir repo/yliss/litlib
+	[[ -L repo/yliss/litlib ]] || rmdir repo/yliss/litlib
 	ln -s $(CURDIR)/repo/litlib repo/yliss/litlib
 	cd repo/yliss && make init
+
+pull:
+	cd repo/yliss && git checkout master && git pull
+	cd repo/litlib && git checkout main && git pull
+
+copy-assets:
+	./script/copy-assets.bash
+
+update: pull copy-assets all
